pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lott"
version = "0.1.0"
description = "LOTT - Libre Open-source Trading Terminal"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "LOTT Team"}
]
keywords = ["trading", "quant", "futures", "backtest", "ctp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    # Core
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    
    # Database
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "redis>=4.0.0",
    
    # Data processing
    "pyarrow>=8.0.0",
    "holidays>=0.18.0",
    
    # Visualization
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "PyQt5>=5.15.0",
    "pyqtgraph>=0.12.0",
    
    # Backtest frameworks
    "backtrader>=1.9.0",
    "vectorbt>=0.25.0",
    "backtesting>=0.3.0",
    
    # Statistics & ML
    "statsmodels>=0.12.0",
    "scikit-learn>=1.0.0",
    
    # Utils
    "python-dateutil>=2.8.0",
    "pytz>=2021.0",
    "tqdm>=4.60.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=3.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.15.0",
]

ctp = [
    # CTP API dependencies (requires manual installation)
    # vnpy-ctp>=2.0.0
]

timescaledb = [
    "timescaledb>=0.1.0",
]

all = [
    "lott[dev,ctp,timescaledb]",
]

[project.urls]
Homepage = "https://github.com/lott-project/lott"
Documentation = "https://github.com/lott-project/lott/docs"
Repository = "https://github.com/lott-project/lott"

[project.scripts]
lott = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "src/*"]
namespaces = true

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml", "*.ini"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | venv
    | conda
    | ref
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".hg", ".mypy_cache", ".tox", ".venv", "_build", "buck-out", "build", "dist", "venv", "conda", "ref"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    "conda",
    "ref",
    "tests",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/conda/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]